requirements.txt
Flask==3.0.0
pytesseract==0.3.10
pdfplumber==0.10.3
Pillow==10.1.0
google-generativeai==0.7.2
gunicorn==21.2.0